+++ modules/dummy_in/dummy_in.c
@@ -354,7 +354,7 @@ static GF_Descriptor *DC_GetServiceDesc(GF_InputService *plug, u32 expect_type, 
 	}
 	bs = gf_bs_new(NULL, 0, GF_BITSTREAM_WRITE);
 	gf_bs_write_u32(bs, size);
-	gf_bs_write_data(bs, uri, strlen(uri));
+	gf_bs_write_data(bs, uri, (u32) strlen(uri));
 	gf_bs_get_content(bs, &esd->decoderConfig->decoderSpecificInfo->data, &esd->decoderConfig->decoderSpecificInfo->dataLength);
 	gf_bs_del(bs);
